+++ cpp/src/join/distinct_hash_join.cu
@@ -14,13 +14,12 @@
  * limitations under the License.
  */
 #include "join_common_utils.cuh"
-#include "join_common_utils.hpp"
 
 #include <cudf/detail/cuco_helpers.hpp>
 #include <cudf/detail/join/distinct_hash_join.cuh>
 #include <cudf/detail/nvtx/ranges.hpp>
 #include <cudf/detail/utilities/cuda.cuh>
-#include <cudf/join.hpp>
+#include <cudf/join/hash_join.hpp>
 #include <cudf/table/experimental/row_operators.cuh>
 #include <cudf/table/table_view.hpp>
 #include <cudf/types.hpp>
@@ -38,7 +37,6 @@
 #include <thrust/iterator/zip_iterator.h>
 #include <thrust/sequence.h>
 
-#include <cstddef>
 #include <limits>
 #include <memory>
 #include <utility>
