+++ cpp/src/join/join.cu
@@ -18,7 +18,8 @@
 #include <cudf/detail/gather.cuh>
 #include <cudf/detail/nvtx/ranges.hpp>
 #include <cudf/dictionary/detail/update_keys.hpp>
-#include <cudf/join.hpp>
+#include <cudf/join/join.hpp>
+#include <cudf/join/hash_join.hpp>
 #include <cudf/table/table.hpp>
 #include <cudf/table/table_view.hpp>
 #include <cudf/utilities/memory_resource.hpp>
